Drop old events after some time
The matter of the fact is that some events are stuck in the queue and we
just can't post tweets for them. As discussed earlier this may be
because someone revoked permissions for our app or somehow ran out of
their personal API limit - the exact reason is unclear and doesn't
really matter.

This commit removes events which are stuck in the queue after some time.

The reasoning for this is twofold. Firstly the likelihood of actually
successfuly posting a tweet which is stuck is very low. As far as I can
tell they are just stuck. Secondly posting tweets for notes which are
old can be confusing and I'd personally be surprised if I suddently had
tweets for notes that are weeks old posted in my profile.

This rollout has to happen in two stages. First we start emitting tweet
created events containing new data (nostr events). After a week passes
we drop all tweet created events that don't have nostr events in them
and change the code to no longer support tweet created events with
missing nostr events.